requirements.txt
python-dotenv
requests
google-cloud-bigquery
google-cloud-storage
google-api-core